refactor: move injectQuery/removeQuery to moduleRunner/utils.ts
