+++ packages/vitest/src/runtime/moduleRunner/startVitestModuleRunner.ts
@@ -9,7 +9,8 @@ import { isBareImport } from '@vitest/utils/helpers'
 import { isBrowserExternal, isBuiltin, toBuiltin } from '../../utils/modules'
 import { getSafeWorkerState } from '../utils'
 import { getCachedVitestImport } from './cachedResolver'
-import { removeQuery, unwrapId, VitestModuleEvaluator } from './moduleEvaluator'
+import { unwrapId, VitestModuleEvaluator } from './moduleEvaluator'
+import { removeQuery } from './utils'
 import { VitestMocker } from './moduleMocker'
 import { VitestModuleRunner } from './moduleRunner'
